% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.ypbp}
\alias{coef.ypbp}
\title{Estimated regression coefficients}
\usage{
\method{coef}{ypbp}(object, ...)
}
\arguments{
\item{object}{an object of the class ypbp.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the estimated regression coefficients.
}
\description{
This function returns the estimated regression coefficients when the maximum likelihood estimation approach is used in the model fitting.
}
\examples{
\donttest{
fit <- ypbp(Surv(time, status)~arm, data=ipass)
coef(fit)
}

}
