% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ypbp.R
\name{crossTime}
\alias{crossTime}
\title{Generic S3 method crossTime}
\usage{
crossTime(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the crossing survival time
}
\description{
Generic S3 method crossTime
}
