% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gastric}
\alias{gastric}
\title{Gastric cancer data set}
\format{
A data frame with 90 rows and 3 variables:
\itemize{
  \item time: survival times (in days)
  \item status: failure indicator (1 - failure; 0 - otherwise)
  \item trt: treatments (1 - chemotherapy + radiation; 0 - chemotherapy alone)
}
}
\description{
Data set from a clinical trial conducted by the Gastrointestinal Tumor Study Group (GTSG) in 1982. The data set refers to the survival times of patients with locally nonresectable gastric cancer. Patients were either treated with chemotherapy combined with radiation or chemotherapy alone.
}
\references{
Gastrointestinal Tumor Study Group. (1982) A Comparison of Combination Chemotherapy and Combined Modality Therapy for Locally Advanced Gastric Carcinoma. Cancer 49:1771-7.
}
\keyword{datasets}
