% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.matrix.ypbp}
\alias{model.matrix.ypbp}
\title{Model.matrix method for ypbp models}
\usage{
\method{model.matrix}{ypbp}(object, ...)
}
\arguments{
\item{object}{an object of the class ypbp.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model matrix (or matrices) for the fit.
}
\description{
Reconstruct the model matrix (or matrices if the alternative formulation of the YP model is used) for a ypbp model.
}
\examples{
\donttest{
fit <- ypbp(Surv(time, status)~arm, data=ipass)
model.matrix(fit)
}

}
