% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ypbp.R
\name{print.summary.ypbp}
\alias{print.summary.ypbp}
\title{Print the summary.ypbp output}
\usage{
\method{print}{summary.ypbp}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.ypbp.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.ypbp output
}
