% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ypbp.R
\name{summary.ypbp}
\alias{summary.ypbp}
\title{Summary for the ypbp model}
\usage{
\method{summary}{ypbp}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'ypbp'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the ypbp model
}
