% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.ypbp}
\alias{vcov.ypbp}
\title{Variance-covariance matrix for a ypbp model}
\usage{
\method{vcov}{ypbp}(object, ...)
}
\arguments{
\item{object}{an object of the class ypbp.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
This function extracts and returns the variance-covariance matrix associated with the regression coefficients when the maximum likelihood estimation approach is used in the model fitting.
}
